Introduce support for plugins: Part 3: build.js
Introduce a simple build system that takes the core
impress.js file and all files and concatenates them
into js/impress.js.

Note that the build script is based on Node for platform
independence. Running impress.js in a browser doesn't depend
on Node or anything else outside of this repository.