Add plugin extras which will initialize extras/ addons for you.
To make loading of extras/ addons simpler, and remove cruft from
presentation html files, created a new plugin src/plugins/extras/.

If any of the extras/impress-console, extras/mermaid/,
extras/highlight or extras/markdown (see next commit) addons
are added to the html file with a regular <script> tag, then
this module will discover and know how to init the module for you,
so it is not necessary to do that in html. If you're not using
the extras, this plugin does nothing.