Add mouse-timeout plugin and change toolbar to depend on it.
This commit adds a generic mouse-timeout plugin and removes the same
functionality from inside the toolbar plugin. As a result, we have
less code and less CSS needed on the user side as well. Feel pretty
smug about myself now :-)

Although this implementation is different and more generic, the
suggestion to add ability to hide mouse cursor came from
a pull request by Sebastian Clausen (@sclausen):
https://github.com/impress/impress.js/pull/536

Description from README:

After 3 seconds of mouse inactivity, add the css class
`body.impress-mouse-timeout`. On `mousemove`, `click` or `touch`, remove the
class.

The use case for this plugin is to use CSS to hide elements from the screen
and only make them visible when the mouse is moved. Examples where this
might be used are: the toolbar from the toolbar plugin, and the mouse cursor
itself.

Example CSS

    body.impress-mouse-timeout {
        cursor: none;
    }
    body.impress-mouse-timeout div#impress-toolbar {
        display: none;
    }