Go to prev() step when shift+tab is pressed.
The tab key is used for navigation to override the browser default
behavior: navigating to next link, button or other widget, which
would cause the browser to scroll to that widget being visible
in the window.

However, since some time ago, shift+tab (and shift+anything)
would be ignored. Thus, if user presses shift+tab, the unwanted
default behavior still happens. (...just in the opposite direction.)
This patch adds shift+tab as a special case, and ties it to
impress().prev(). Anything else with shift pressed will still be
ignored.

References:
https://github.com/impress/impress.js/issues/116
https://github.com/impress/impress.js/pull/550