Only execute preStepLeave plugins from goto, when the new step is different from activeStep.
...in other words, when we really are about to leave the active step.
This is analogous to how the impress:stepleave event works as well.

This also changes back a unit test, which caught this error when I introduced this,
but didn't realize it at the time, as there were other unrelated race conditions in the tests.

Fixes 09ad1d82b7f79cbb7ec14c35a4c146a97dc43465