Fix the z-axis for 2 steps in the official demo. (Needs explicit 0.)
The relative positioning plugin (rel.js) includes a small (and
intentional) backward incompatible change: Previously a missing
data-x/y/z attribute would have the default value of 0. However,
the rel.js plugin causes missing values to be inherited from the
previous step instead.

In most cases this shouldn't be an issue, as it wouldn't make
sense for legacy presentations to not specify coordinates
for steps. However, it turns out it can happen more often for
the z-axis: steps deviating for the xy-plane have a data-z
value, but steps in the xy-plane simply omit the data-z
attribute, instead of explicitly specifying 0.

In fact, embarassingly, this was also the case for the official
demo (/index.html) in impress.js repo itself! This commit adds
two explicit data-z="0" where necessary.

Realizing that the backward incompatibility might be a bit more
annoying than anticipated, also added a note to
src/plugins/rel/README.md.

(Part of this work was to also consider reducing or removing the
backward incompatibility, however I chose to keep it this way.
While backward compatibility is important, in this case I'd rather
be inconsistent with the past, than carry the inconsistency with
me into the future.)