Move addPreInitPlugin() and addPreStepLeavePlugin() to be methods of the impress object.
I realized that even though window.impress() is a function, which
returns an object of more functions btw, it is also possible for
window.impress to have member variables and methods, because in JS
everything is an object. For the pre-init and pre-stepleave plugins
this is more appropriate, so this commit moves them from the api
to being truly global (and singleton) methods.

Old:
impress().addPreInitPlugin()

New:
impress.addPreInitPlugin()