Show substeps also in impress console window
The slide view in the impress console window is actually an independent
presentation. The movements to previous and next slides are simply
propagated from the main presenataion by listening to the impress events
and then replicating those to the slide-view and preview.

Same must be done for substeps. This adds new events impress:substep:show
and impress:substep:hide that can be used to show/hide the next substep.
These commands are then used by impressConsole.js to keep the slide view
in sync with the main window.